feat(recommendation): add webExtendEvent to request mapping ticket: SDK-1302 epic: SDK-1209

+++ engagement-cloud-sdk/src/commonMain/kotlin/com/sap/ec/webExtend/CartItem.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CartItem(
     val itemId: String,
-    val price: Float,
-    val quantity: Int,
+    val price: Double,
+    val quantity: Double,
 )
